openssl: BN_set_params 1U<< after the clamp to 31, and first-party arithmetic is closed
Same type-level 1<<31 as IFCAP_BIT. The remaining bin/sbin/usr.bin/stand/crypto arithmetic queue is in arithmetic-triage.json; FuSeBMC NOSEED is missing contrib test headers, not a proof gap to fuzz harder.

+++ hbsd/src/crypto/openssl/crypto/bn/bn_lib.c
@@ -42,25 +42,25 @@ void BN_set_params(int mult, int high, int low, int mont)
         if (mult > (int)(sizeof(int) * 8) - 1)
             mult = sizeof(int) * 8 - 1;
         bn_limit_bits = mult;
-        bn_limit_num = 1 << mult;
+        bn_limit_num = (int)(1U << mult);
     }
     if (high >= 0) {
         if (high > (int)(sizeof(int) * 8) - 1)
             high = sizeof(int) * 8 - 1;
         bn_limit_bits_high = high;
-        bn_limit_num_high = 1 << high;
+        bn_limit_num_high = (int)(1U << high);
     }
     if (low >= 0) {
         if (low > (int)(sizeof(int) * 8) - 1)
             low = sizeof(int) * 8 - 1;
         bn_limit_bits_low = low;
-        bn_limit_num_low = 1 << low;
+        bn_limit_num_low = (int)(1U << low);
     }
     if (mont >= 0) {
         if (mont > (int)(sizeof(int) * 8) - 1)
             mont = sizeof(int) * 8 - 1;
         bn_limit_bits_mont = mont;
-        bn_limit_num_mont = 1 << mont;
+        bn_limit_num_mont = (int)(1U << mont);
     }
 }
